clean-up / suppress warnings / lock orientation to landscape / prevent image acquisition from crashing App (disabled)

+++ cmEditDemo/cmEditDemo/Homepwner/BNRItemStore.h
@@ -28,8 +28,8 @@
 
 - (BNRItem *)createItem;
 
-- (void)moveItemAtIndex:(int)from
-                toIndex:(int)to;
+- (void)moveItemAtIndex:(NSInteger)from
+                toIndex:(NSInteger)to;
 
 - (NSString *)itemArchivePath;
